+++ pMuTT/mixture/cov.py
@@ -7,7 +7,7 @@ Created on Thurs Feb 7 10:15:00 2019
 
 import numpy as np
 from pMuTT import constants as c
-from pMuTT.io.jsonio import remove_class
+from pMuTT.io.json import remove_class
 
 
 class CovEffect:
