+++ pMuTT/statmech/trans.py
@@ -3,7 +3,7 @@
 import numpy as np
 from pMuTT import constants as c
 from pMuTT import get_molecular_weight
-from pMuTT.io.jsonio import remove_class
+from pMuTT.io.json import remove_class
 
 
 class IdealTrans:
