+++ pMuTT/statmech/vib.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 from pMuTT import constants as c
-from pMuTT.io.jsonio import remove_class
+from pMuTT.io.json import remove_class
 
 
 class HarmonicVib:
